requirements.txt
pymongo==4.3.3
motor==3.1.2
dnspython==2.3.0
pydantic==1.10.7
fastapi==0.95.1
python-dotenv==1.0.0
pandas==2.0.0
matplotlib==3.8.0
python-dateutil==2.8.2
spacy==3.7.1
SQLAlchemy==1.3.24
mathparse==0.1.2
pytz==2023.3.post1
numpy==1.26.0
scipy==1.11.3
scikit-learn==1.3.1
nltk==3.8.1
networkx==3.1
geopy==2.4.0
uvicorn
langdetect==1.0.9